+++ src/lists/typeDefs.js
@@ -1,19 +1,19 @@
 export const listsTypeDef = `
 type List {
-    code: Long!
+    code: Int!
     name: String!
     archived: Boolean!
     description: String!
     created_at: Date!
     updated_at: Date!
-    board_code: Long!
+    board_code: Int!
 
 }
 input ListInput {
     name: String!
     archived: Boolean!
     description: String!
-    board_code: Long!
+    board_code: Int!
 }`;
 
 export const listsQueries = `
